Add option to install a local package to be deployed as well
If you want to deploy your lambda function and need a dependency which
is not available on PyPi, but only to your local machine (private package, ...)
you can now add this to the dist archive as well.

For both, the `build` and `deploy` option, an optional argument `--local-package`
can be provided with the path to the package you want to install via pip.
Your package needs to have a `setup.py` file.

Example:
    $ lambda build --local-package .