Composing path to bootstrap directory should be done with posixpath join
Most of the people probably bootstrap the project giving no or just "basename"
folder to "lambda init" command, but if users would give full path as a folder,
the project will be created into working directory, but into several sub-directories.

So, not even a issue in most cases, but os.path.join would make it bit straight-forward?

Example:

	$ # current working directory
	$ pwd
	/Users/tanel/tmp/lambda
	$
	$ # no files of directories here
	$ tree
	.

	0 directories, 0 files
	$
	$ # just to see how bash replaces the ~
	$ echo lambda init ~/tmp/parent
	lambda init /Users/tanel/tmp/parent
	$
	$ # and now, lets executing it
	$ lambda init ~/tmp/parent
	$
	$ # result - bit odd sub-directory structure in working directory
	$ tree
	.
	└── Users
	    └── tanel
	        └── tmp
	            └── parent
	                ├── config.yaml
	                ├── event.json
	                └── service.py

	4 directories, 3 files