Related to issue #64 "Subdirectories not included in zip package".
Added the ability to explicitly define the list of "source" subdirectories in the config.yaml file under a new "build" configuration section.  Decided it was perhaps best to allow explicit declaration of the folder list so we don't build undesirable (.git) folders into the zipfile.  Modified the template config.yaml with this setting as well.