Idea/Tentative Change: Don't set role by default
If the user wants to manage the role through the AWS web interface, or doesn't want to give the user updating the function the iam:PassRole permission for security, don't change the role back to lambda_basic_execution.  If the role is not defined in the config, it will not be updated on Lambda.  If more checks like this are implemented they might be moved to a dedicated part of the code.